Added a pause menu. Changed menus to now have a boolean for whether they can be clicked or not. Made Button class accept a Lambda for programming its function.
