+++ src/main/java/org/bcit/com2522/project/Button.java
@@ -1,7 +1,5 @@
 package org.bcit.com2522.project;
 
-import processing.core.PVector;
-
 import java.awt.*;
 
 public class Button {
@@ -9,13 +7,13 @@ public class Button {
   float y;
   float width;
   float height;
-  PVector position;
   Color bg;
   String label;
   Window window;
-  AltMenu menu;
+  Menu menu;
+  private Executable function;
 
-  public Button(String s, float initX, float initY, float w, float h, Color c, Window space, AltMenu m){
+  public Button(String s, float initX, float initY, float w, float h, Color c, Window space, Menu m){
     width = w;
     height = h;
     x = initX;
@@ -24,6 +22,10 @@ public class Button {
     label = s;
     window = space;
     menu = m;
+  }
+
+  public void config(Executable ex){
+    function = ex;
   }
 
   public void draw(){
@@ -43,6 +45,11 @@ public class Button {
   }
 
   public void execute() {
-    window.initializeObjects();
+    function.execute();
+    menu.hideMenu();
   }
+}
+
+interface Executable{
+  void execute();
 }
