+++ src/main/java/org/bcit/com2522/project/Window.java
@@ -12,6 +12,8 @@ import processing.core.PVector;
 import processing.event.KeyEvent;
 import processing.event.MouseEvent;
 
+import java.util.ArrayList;
+
 
 /**
  * The window class runs the display that initializes and displays all
@@ -51,7 +53,8 @@ public class Window extends PApplet {
     LOAD,
     GAMEOVER,
     PLAY,
-    WIN
+    WIN,
+    PAUSE
   }
 
   String funFact;
@@ -62,7 +65,9 @@ public class Window extends PApplet {
   float playerAnimationTime;
   float elpCount = 0;
   int loadingTimer = 0;
-  AltMenu menu;
+  MainMenu menu = new MainMenu(this);
+  PauseMenu pauseMenu = new PauseMenu(this);
+  ArrayList<Menu> menus = new ArrayList<Menu>();
 
   /**
    * Provides the size of the window
@@ -80,6 +85,9 @@ public class Window extends PApplet {
     minim = new Minim(this);
     sound = minim.loadFile("sound/heroSong.mp3");
 
+    menus.add(menu);
+    menus.add(pauseMenu);
+
     frameRate(FPS);
 
     //sets up the background image
@@ -88,7 +96,7 @@ public class Window extends PApplet {
     // initializes the objects
     this.initializeObjects();
     state = State.MENU;
-    menu = new AltMenu(this);
+    menu.loadMenu();
   }
 
 
@@ -123,7 +131,6 @@ public class Window extends PApplet {
     playerAnimationTime = timer.getTime();
     switch( keyCode ) {
       case LEFT:
-        // handle left
         player.setDirection(new PVector(-2, 0));
         if (Math.round(playerAnimationTime*5) % 2 == 0){
           player.setHarryPotterImage(player.playerLeftWalk1);
@@ -132,7 +139,6 @@ public class Window extends PApplet {
         }
         break;
       case RIGHT:
-        // handle right
         player.setDirection(new PVector(2, 0));
         if (Math.round(playerAnimationTime*5) % 2 == 0){
           player.setHarryPotterImage(player.playerRightWalk1);
@@ -141,7 +147,6 @@ public class Window extends PApplet {
         }
         break;
       case UP:
-        // handle left
         player.setDirection(new PVector(0, -2));
         if (Math.round(playerAnimationTime*5) % 2 == 0){
           player.setHarryPotterImage(player.playerUp1);
@@ -150,7 +155,6 @@ public class Window extends PApplet {
         }
         break;
       case DOWN:
-        // handle right
         player.setDirection(new PVector(0, 2));
         if (Math.round(playerAnimationTime*5) % 2 == 0){
           player.setHarryPotterImage(player.playerDown1);
@@ -199,13 +203,17 @@ public class Window extends PApplet {
       case 'M':
         if (state == State.WIN || state == State.GAMEOVER);
         state = state.MENU;
+        menu.loadMenu();
         break;
-
+      case 'P':
+        if (state == State.PLAY){
+          state = state.PAUSE;
+        }
     }
   }
 
   public void mouseClicked(MouseEvent m){
-    if (state == State.MENU){
+    for (Menu menu : menus){
       menu.click(m);
     }
   }
@@ -219,6 +227,7 @@ public class Window extends PApplet {
 
     switch (state) {
       case MENU:
+        menu.loadMenu();
         menu.draw();
         break;
       case LOAD:
@@ -248,7 +257,6 @@ public class Window extends PApplet {
         break;
 
       case PLAY:
-
         if (timer == null) {
           timer = new Timer(this, new PVector(0, 0));
         }
@@ -306,6 +314,7 @@ public class Window extends PApplet {
 
       case GAMEOVER:
         background(0);
+        fill(255);
         textSize(50);
         text("Game Over!", width / 3, height / 2);
         text("Press R to restart.", width / 3, height / 2 + 50);
@@ -317,11 +326,19 @@ public class Window extends PApplet {
         sound.pause();
         background(0);
         textSize(50);
+        fill(255);
         text("You Won!!!", width / 3, height / 2);
         textSize(30);
         String time = String.format("%.1f", timeElapsed);
         text("Your time was " + time + " seconds!", width / 4, height / 2 + 50);
         text("Press M to return to menu", width / 3, height / 2 + 100);
+        break;
+
+      case PAUSE:
+        sound.pause();
+        pauseMenu.loadMenu();
+        pauseMenu.draw();
+        break;
     }
   }
